requirements.txt
Django==3.1.4
django-admin-interface
django-annoying==0.10.6
django-colorfield==0.4.1
django-cryptography==1.0
django-extensions==3.1.0
django-model-utils==4.1.1
django-minio-storage==0.3.10
django-redis==4.12.1
django-redis-sessions==0.6.2
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
minio==6.0.2
ipdb
psycopg2-binary==2.8.6
python-jose==3.2.0
redis==3.5.3
uWSGI==2.0.18
pytz~=2020.5
Pillow==8.2.0
whitenoise==5.2.0
cryptography~=3.4.7